requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
requests==2.31.0
pyromod==1.5
ffmpeg
mediainfo
cloudscraper
beautifulsoup4==4.9.3
bs4==0.0.1
aiohttp==3.8.4
aiofiles
pycryptodome==3.10.1
PySocks==1.7.1
websockets==11.0.3
yt-dlp==2023.6.22
psutil
async-lru==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
mutagen==1.46.0
helper
pyaes==1.6.1
soupsieve==2.4.1
py-tgcalls==0.6.0
jinja2
urllib3==2.0.3
pytz
umongo==3.1.0
speedtest-cli